+++ run_configs/lfric/lfric_common.py
@@ -4,11 +4,8 @@ import shutil
 from pathlib import Path
 from typing import Dict
 
-from fab.artefacts import SuffixFilter
-from fab.steps import Step, check_for_errors
-from fab.steps.preprocess import PreProcessor
+from fab.steps import Step
 from fab.tools import run_command
-from fab.util import log_or_dot, input_to_output_fpath
 
 logger = logging.getLogger('fab')
