+++ source/fab/steps/analyse.py
@@ -41,12 +41,12 @@ from pathlib import Path
 from typing import Dict, List, Iterable, Set, Optional, Union
 
 from fab.artefacts import ArtefactsGetter, CollectionConcat, SuffixFilter
-from fab.constants import BUILD_TREES, CURRENT_PREBUILDS
+from fab.constants import BUILD_TREES
 from fab.dep_tree import add_mo_commented_file_deps, extract_sub_tree, \
     validate_dependencies
 from fab.parse import AnalysedFile, EmptySourceFile, AnalysedDependent
 from fab.parse.c import CAnalyser
-from fab.parse.fortran.fortran import FortranParserWorkaround, FortranAnalyser
+from fab.parse.fortran import FortranParserWorkaround, FortranAnalyser
 from fab.steps import Step
 from fab.util import TimerLogger, by_type
